moved terms showmore to component
